web/requirements.txt
asgiref==3.7.2
Django==3.2.6
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
Markdown==3.4.4
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
